Support to pack binary from local source

+++ infra/pack_bin.sh
@@ -6,34 +6,38 @@ SOURCE_ROOT="${ROOT_PARENT}/${SOURCE_FOLDER_NAME}"
 TODAY=$(date +"%Y%m%d")
 NF_PATH="${SOURCE_ROOT}/NFs"
 
-if [ "${1}" == "" ]; then
-    BIN_FOLDER_NAME="bin_${TODAY}"
-    git clone https://bitbucket.org/free5gc-team/free5gc.git ${SOURCE_ROOT}
-    cd ${SOURCE_ROOT} && git checkout develop && git submodule sync --recursive && git submodule update --init && git submodule foreach git checkout develop && git submodule foreach git pull
+if [ "${1}" == "local" ]; then
+        BIN_FOLDER_NAME="bin_local"
 else
-    BIN_FOLDER_NAME="bin_${1}"
-    git clone --recursive -b ${1} https://bitbucket.org/free5gc-team/free5gc.git ${SOURCE_ROOT}
-    cd ${SOURCE_ROOT}
+    if [ "${1}" == "" ]; then
+        BIN_FOLDER_NAME="bin_${TODAY}"
+        git clone https://bitbucket.org/free5gc-team/free5gc.git ${SOURCE_ROOT}
+        cd ${SOURCE_ROOT} && git checkout develop && git submodule sync --recursive && git submodule update --init && git submodule foreach git checkout develop && git submodule foreach git pull
+    else
+        BIN_FOLDER_NAME="bin_${1}"
+        git clone --recursive -b ${1} https://bitbucket.org/free5gc-team/free5gc.git ${SOURCE_ROOT}
+        cd ${SOURCE_ROOT}
+    fi
+
+    # go vendor
+    # NF
+    for NF in ${NF_PATH}/*
+    do
+        cd ${NF}
+        go mod vendor
+        cd ~-
+    done
+
+    # webconsole
+    cd ${SOURCE_ROOT}/webconsole
+    go mod vendor
+    cd ~-
 fi
+
+make all
 
 BIN_ROOT="${ROOT_PARENT}/${BIN_FOLDER_NAME}"
 TGZ_FILE_NAME="${BIN_FOLDER_NAME}.tgz"
-
-# go vendor
-# NF
-for NF in ${NF_PATH}/*
-do
-    cd ${NF}
-    go mod vendor
-    cd ~-
-done
-
-# webconsole
-cd ${SOURCE_ROOT}/webconsole
-go mod vendor
-cd ~-
-
-make all
 
 echo "Pack the binary..."
 mkdir -p ${BIN_ROOT}/NFs/upf/build/utlt_logger
@@ -50,8 +54,16 @@ cp -rf webconsole/config ${BIN_ROOT}/webconsole
 cp -rf webconsole/public ${BIN_ROOT}/webconsole
 cp -rf run.sh ${BIN_ROOT}
 cp -rf force_kill.sh ${BIN_ROOT}
-cd ..
+
+if [ "${1}" != "local" ]; then
+    cd ..
+fi
+
 tar zcf ${TGZ_FILE_NAME} ${BIN_FOLDER_NAME}
-rm -rf ${SOURCE_FOLDER_NAME}
+
+if [ "${1}" != "local" ]; then
+    rm -rf ${SOURCE_FOLDER_NAME}
+fi
+
 rm -rf ${BIN_FOLDER_NAME}
 echo "Pack done"
